+++ packages/docusaurus-utils-validation/src/index.ts
@@ -29,6 +29,4 @@ export {
   FrontMatterAuthorErrorMessage,
   FrontMatterLastUpdateErrorMessage,
   FrontMatterLastUpdateSchema,
-  validateBlogPostFrontMatter,
-  validateDocFrontMatter,
 } from './validationSchemas';
